backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "boto3>=1.42.90",
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "resend>=2.29.0",
    "solders>=0.27.1",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
